requirements.txt
requests
playsound
aiohttp
aiofiles
vosk
pygame
pyaudio
selenium>=4.9.0
webdriver-manager>=3.8.6
pvporcupine
python-dotenv
opencv-python
pillow 
mss
google-genai